pyproject.toml
[tool.poetry]
readme = "README.md"
repository = "https://github.com/isaevmik/hse_meeting_bot"
name = "hse_meeting_bot"
version = "1.1.0"
description = "-"
authors = [
    "Mikhail Isaev <warlockmikhail@gmail.com>",
]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.8.3"
aiosignal = "1.3.1"
apscheduler = "3.6.3"
async-timeout = "4.0.2"
attrs = "22.1.0"
black = "22.10.0"
cachetools = "4.2.2"
captcha = "0.4"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
click = "8.1.3"
ez-setup = "0.9"
frozenlist = "1.3.3"
idna = "3.4"
multidict = "6.0.3"
mypy-extensions = "0.4.3"
pathspec = "0.10.2"
peewee = "3.15.4"
pillow = "9.3.0"
platformdirs = "2.6.0"
pytelegrambotapi = "4.8.0"
python-decouple = "3.6"
python-dotenv = "0.21.0"
python-telegram-bot = "13.15"
pytz = "2022.6"
pytz-deprecation-shim = "0.1.0.post0"
requests = "2.28.1"
six = "1.16.0"
tomli = "2.0.1"
tornado = "6.1"
tzdata = "2022.7"
tzlocal = "4.2"
urllib3 = "1.26.13"
yarl = "1.8.2"
psycopg2-binary = "2.9.5"
pylint = "^2.15.10"
flake8 = "^6.0.0"
mypy = "^0.991"
autoflake = "^2.0.0"
isort = "^5.11.4"
unify = "^0.5"
